+++ src/frontend/environments/environment.beta.ts
@@ -9,6 +9,7 @@ export const environment: AppEnvironment = {
     tracksBaseUrl: 'http://localhost:4101/api/assets/trackstore/',
     audioBaseUrl: 'http://localhost:4101/api/static/audiostore/',
     picsBaseUrl: 'http://localhost:4101/api/static/picturestore/',
+    picsPreviewPathResolution: 'x300',
     videoBaseUrl: 'http://localhost:4101/api/static/videostore/',
     defaultSearchTypes: 'route,location,track,trip,news',
     emptyDefaultSearchTypes: 'route,location,track,trip,news,image,video,info',