+++ docs/CHANGELOG.md
@@ -34,6 +34,7 @@
 ### breaking changes
 - build: upgraded all dev-dependencies to latest
 - build: use peerDependencies
+- frontend: configure preview-resolution in environment
 - common: used x300 instead of x600 for image-preview -> run scaleImages to create non-existing images
 ```
 run node dist/backend/serverAdmin.js --command imageManager --action scaleImages -c config/backend.json --debug --parallel 10